Backend/requirements.txt
streamlit
pymongo
pymupdf
pandas
fpdf2
certifi